Refactor telemetry metrics handling: replace MetricSnapshot with JobMetrics
- Updated TelemetryServiceExtensions to register IJobMetricsStore instead of IMetricSnapshotStore.
- Modified AgentControlPlaneClientAdapter to return JobMetrics in GetTelemetryAsync.
- Refactored ControlPlaneTelemetryTimer to push JobMetrics to the Control Plane.
- Adjusted MigrationAgentServiceExtensions to reflect changes in telemetry metrics.
- Updated TUI tests to handle JobMetrics instead of MetricSnapshot.
- Refined JobProgressStoreCurrentStateTests to work with new event structure.
- Enhanced DependencyDiscoveryModuleResumeTests to validate event emissions.
- Revised WorkItemExportOrchestratorTests to ensure progress events reflect JobMetrics.
- Updated ControlPlaneTelemetryClientTests to push JobMetrics and validate request bodies.
- Modified MetricSnapshotSerializationTests to serialize and deserialize JobMetrics correctly.
- Refactored SnapshotMetricExporterTests to test InMemoryJobMetricsStore functionality.

+++ src/DevOpsMigrationPlatform.Abstractions/Models/InventorySummary.cs
@@ -10,10 +10,10 @@ public sealed class InventorySummary
 {
     public string Url { get; set; } = string.Empty;
     public string ProjectName { get; set; } = string.Empty;
-    public int WorkItemsCount { get; set; }
-    public int RevisionsCount { get; set; }
-    public int ReposCount { get; set; }
-    public int PipelinesCount { get; set; }
+    public long WorkItemsCount { get; set; }
+    public long RevisionsCount { get; set; }
+    public long ReposCount { get; set; }
+    public long PipelinesCount { get; set; }
     /// <summary>True when all date windows were scanned without error.</summary>
     public bool IsComplete { get; set; }
     /// <summary>Non-null when counting failed for this project.</summary>
